Bookmarks Manager is created and lastfolder method is switched to use this instance
(cherry picked from commit c0679a473e5607b387a10b96b5539f76da9e1ae2)
