+++ Client/Frontend/Sync/BraveCore/Bookmarks/Bookmarkv2.swift
@@ -17,6 +17,7 @@ private let log = Logger.browserLogger
 class Bookmarkv2: WebsitePresentable {
     private let bookmarkNode: BookmarkNode
     private var bookmarkFavIconObserver: BookmarkModelListener?
+    
     private static var bookmarkModelLoadedObserver: BookmarkModelListener?
     private static let bookmarksAPI = (UIApplication.shared.delegate as? AppDelegate)?.braveCore?.bookmarksAPI
 
@@ -114,76 +115,6 @@ class Bookmarkv2: WebsitePresentable {
         }
     }
     
-    // Returns the last visited folder
-    // If no folder was visited, returns the mobile bookmarks folder
-    // If the root folder was visited, returns nil
-    public static func lastVisitedFolder() -> Bookmarkv2? {
-        guard let bookmarksAPI = self.bookmarksAPI else {
-            return nil
-        }
-        
-        guard Preferences.General.showLastVisitedBookmarksFolder.value,
-              let nodeId = Preferences.Chromium.lastBookmarksFolderNodeId.value else {
-            // Default folder is the mobile node..
-            if let mobileNode = bookmarksAPI.mobileNode {
-                return Bookmarkv2(mobileNode)
-            }
-            return nil
-        }
-        
-        // Display root folder instead of mobile node..
-        if nodeId == -1 {
-            return nil
-        }
-        
-        // Display last visited folder..
-        if let folderNode = bookmarksAPI.getNodeById(nodeId),
-           folderNode.isVisible {
-            return Bookmarkv2(folderNode)
-        }
-        
-        // Default folder is the mobile node..
-        if let mobileNode = bookmarksAPI.mobileNode {
-            return Bookmarkv2(mobileNode)
-        }
-        return nil
-    }
-    
-    public static func lastFolderPath() -> [Bookmarkv2] {
-        guard let bookmarksAPI = Bookmarkv2.bookmarksAPI else {
-            return []
-        }
-        
-        if Preferences.General.showLastVisitedBookmarksFolder.value,
-           let nodeId = Preferences.Chromium.lastBookmarksFolderNodeId.value,
-           var folderNode = bookmarksAPI.getNodeById(nodeId),
-           folderNode.isVisible {
-            
-            // We don't ever display the root node
-            // It is the mother of all nodes
-            let rootNodeGuid = bookmarksAPI.rootNode?.guid
-            
-            var nodes = [BookmarkNode]()
-            nodes.append(folderNode)
-            
-            while true {
-                if let parent = folderNode.parent, parent.isVisible, parent.guid != rootNodeGuid {
-                    nodes.append(parent)
-                    folderNode = parent
-                    continue
-                }
-                break
-            }
-            return nodes.map({ Bookmarkv2($0) }).reversed()
-        }
-        
-        // Default folder is the mobile node..
-        if let mobileNode = bookmarksAPI.mobileNode {
-            return [Bookmarkv2(mobileNode)]
-        }
-        
-        return []
-    }
 }
 
 class BraveBookmarkFolder: Bookmarkv2 {
